设置版本为 v1.1.0，支持 Delete Artifacts配置项且兼容 Windows 平台

+++ src/main/java/io/jenkins/plugins/QiniuArtifactManager.java
@@ -16,7 +16,6 @@ import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
 
 import javax.annotation.Nonnull;
-import java.io.File;
 import java.io.IOException;
 import java.io.Serializable;
 import java.util.HashMap;
@@ -43,7 +42,7 @@ public class QiniuArtifactManager extends ArtifactManager {
         this.standardArtifactManager = new StandardArtifactManager(run);
         this.marker = new Marker(this.objectNamePrefixWithBuildNumber, this.config);
         LOG.log(Level.INFO, "QiniuArtifactManager is constructed, accessKey={0}, prefix={1}",
-                new Object[]{this.config.getAccessKey(), this.objectNamePrefixWithBuildNumber});
+                new Object[] { this.config.getAccessKey(), this.objectNamePrefixWithBuildNumber });
     }
 
     @Override
@@ -57,18 +56,19 @@ public class QiniuArtifactManager extends ArtifactManager {
     @Nonnull
     private String generateObjectNamePrefixWithBuildNumber(@Nonnull Run<?, ?> run) {
         String n = this.config.getObjectNamePrefix();
-        if (!n.isEmpty() && !n.endsWith(File.separator)) {
-            n += File.separator;
+        if (!n.isEmpty() && !n.endsWith(QiniuFileSystem.SEPARATOR)) {
+            n += QiniuFileSystem.SEPARATOR;
         }
         n += run.getParent().getFullName();
-        n += File.separator;
+        n += QiniuFileSystem.SEPARATOR;
         n += run.getId();
-        n += File.separator;
+        n += QiniuFileSystem.SEPARATOR;
         return n;
     }
 
     @Override
-    public void archive(FilePath workspace, Launcher launcher, BuildListener buildListener, Map<String, String> pathMap) throws IOException, InterruptedException {
+    public void archive(FilePath workspace, Launcher launcher, BuildListener buildListener, Map<String, String> pathMap)
+            throws IOException, InterruptedException {
         if (!this.config.isApplyForAllJobs() && !this.marker.didUseQiniuArtifactArchiver()) {
             LOG.log(Level.INFO, "StandardArtifactManager::archive()");
             this.standardArtifactManager.archive(workspace, launcher, buildListener, pathMap);
@@ -81,14 +81,17 @@ public class QiniuArtifactManager extends ArtifactManager {
         final Map<String, String> artifacts = new HashMap<>();
         for (Map.Entry<String, String> entry : pathMap.entrySet()) {
             final String objectNameWithoutPrefix = entry.getKey();
-            final String filePath = entry.getValue();
+            final String filePath = QiniuFileSystem.fromFileSystemPathToObjectName(entry.getValue());
             artifacts.put(objectNameWithoutPrefix, filePath);
         }
-        workspace.act(new QiniuUploader(this.config, this.marker, artifacts, this.objectNamePrefixWithBuildNumber, buildListener));
+        workspace.act(new QiniuUploader(this.config, artifacts, this.objectNamePrefixWithBuildNumber, buildListener));
     }
 
     @Override
     public boolean delete() throws IOException, InterruptedException {
+        if (!this.config.isDeleteArtifacts()) {
+            return false;
+        }
         if (!this.marker.didUseQiniuArtifactArchiver()) {
             LOG.log(Level.INFO, "StandardArtifactManager::delete()");
             return this.standardArtifactManager.delete();
@@ -108,7 +111,8 @@ public class QiniuArtifactManager extends ArtifactManager {
             return this.standardArtifactManager.root();
         }
         LOG.log(Level.INFO, "QiniuArtifactManager::root(): prefix={0}", this.objectNamePrefixWithBuildNumber);
-        final QiniuFileSystem qiniuFileSystem = QiniuFileSystem.create(this.config, this.objectNamePrefixWithBuildNumber);
+        final QiniuFileSystem qiniuFileSystem = QiniuFileSystem.create(this.config,
+                this.objectNamePrefixWithBuildNumber);
         return new QiniuFile(qiniuFileSystem, null);
     }
 
@@ -143,9 +147,8 @@ public class QiniuArtifactManager extends ArtifactManager {
         public void useQiniuArtifactArchiver() throws IOException {
             final UploadManager uploadManager = new UploadManager(this.config.getConfiguration());
             final Auth auth = Auth.create(this.config.getAccessKey(), this.config.getSecretKey().getPlainText());
-            final String uploadToken = auth.uploadToken(
-                    this.config.getBucketName(), null,
-                    24 * 3600, new StringMap().put("fileType", 1).put("insertOnly", 0));
+            final String uploadToken = auth.uploadToken(this.config.getBucketName(), null, 24 * 3600,
+                    new StringMap().put("fileType", 1).put("insertOnly", 0));
             Initializer.setAppName();
             uploadManager.put("{}".getBytes("UTF-8"), this.objectName, uploadToken, null, null, true);
         }
@@ -157,7 +160,7 @@ public class QiniuArtifactManager extends ArtifactManager {
         @Nonnull
         private String getMarkObjectName(@Nonnull final String objectNamePrefix) {
             String name = objectNamePrefix;
-            while (name.endsWith(File.separator)) {
+            while (name.endsWith(QiniuFileSystem.SEPARATOR)) {
                 name = name.substring(0, name.length() - 1);
             }
             return name + ".qiniu-artifact-archiver";
